Minmaxlist

Adds min_list_by/2, max_list_by/2 and min_max_list_by/2 that’s not in Enum.

Types

element :: any

t :: Enumerable.t

Functions

max_list_by(enumerable, fun)

Specs

max_list_by(t, (element -> any)) :: list

Returns list of the biggest elements in the enumerable as calculated by the given function.

Returns empty list if the enumerable is empty.

Examples

iex> Minmaxlist.max_list_by(["a", "b", "aa", "bb", "bbb", "aaa"], fn(x) -> String.length(x) end)
["bbb", "aaa"]

iex> Minmaxlist.max_list_by([], fn(x) -> String.length(x) end)
[]

min_list_by(enumerable, fun)

Specs

min_list_by(t, (element -> any)) :: list

Returns list of the smallest elements in the enumerable as calculated by the given function.

Returns empty list if the enumerable is empty.

Examples

iex> Minmaxlist.min_list_by(["a", "b", "aa", "bb", "bbb", "aaa"], fn(x) -> String.length(x) end)
["a", "b"]

iex> Minmaxlist.min_list_by([], fn(x) -> String.length(x) end)
[]

min_max_list_by(enumerable, fun)

Specs

min_max_list_by(t, (element -> any)) :: {list, list}

Returns a tuple with list of smallest elements and list of biggest elements in the enumerable as calculated by the given function.

Returns a tuple of two empty lists if the enumerable is empty.

Examples

iex> Minmaxlist.min_max_list_by(["a", "b", "aa", "bb", "bbb", "aaa"], fn(x) -> String.length(x) end)
{["a", "b"], ["bbb", "aaa"]}

iex> Minmaxlist.min_max_list_by([], fn(x) -> String.length(x) end)
{[], []}